change balance from big integer to decimal to get correct numbers | fix login email error

+++ database/migrations/2026_02_23_151900_create_memberships_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('colocation_id');
             $table->timestamp('left_at')->nullable();
-            $table->bigInteger('balance')->default(0);
+            $table->decimal('balance',10,2)->default(0);
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->enum('role',['owner','member'])->default('member');
             $table->foreign('user_id')->references('id')->on('users');
